refactor: add attribute for favourite in Place and move model to another directory
